moves tests around, adds test for tei contributor
